+++ src/esphome/entities/fan.lua
@@ -32,7 +32,7 @@ function FanEntity:discovered(entity)
     class = class .. "_REVERSE"
   end
   local bindingId = assert(
-    bindings:getOrAddDynamicBinding(self.TYPE, "fan_" .. entity.key, "PROXY", true, entity.name, class)
+    bindings:getOrAddDynamicBinding(self.TYPE, "fan_" .. entity.ref, "PROXY", true, entity.name, class)
   ).bindingId
   RFP[bindingId] = function(idBinding, strCommand, tParams, args)
     log:trace("RFP idBinding=%s strCommand=%s tParams=%s args=%s", idBinding, strCommand, tParams, args)
@@ -43,6 +43,7 @@ function FanEntity:discovered(entity)
         or ESPHomeProtoSchema.RPC.APIConnection.fan_command
       local body = DeserializeSafe(Select(tParams, "body")) or {}
       body.key = body.key or entity.key
+      body.device_id = body.device_id or entity.device_id
       self.client:callServiceMethod(command, body):next(function()
         log:debug(
           "Method %s.%s(%s) called by entity %s",
@@ -81,7 +82,7 @@ end
 --- @return void
 function FanEntity:updated(entity, state)
   log:trace("FanEntity:updated(%s, %s)", entity, state)
-  local binding = bindings:getDynamicBinding(self.TYPE, "fan_" .. entity.key)
+  local binding = bindings:getDynamicBinding(self.TYPE, "fan_" .. entity.ref)
   if binding ~= nil then
     SendToProxy(binding.bindingId, "UPDATE_STATE", {
       entity = SerializeSafe(entity),